go-clarinet update and some other stuff

+++ main.go
@@ -264,7 +264,7 @@ type sampleAppConfig struct {
 	Directory          string
 	NumPeers           int
 	MinPeers           int
-	ActivityPeriodSecs int
+	ActivityPeriodSecs float64
 	TotalActions       int
 }
